login page and player names
