+++ package.json
@@ -15,7 +15,8 @@
     "@fontsource/roboto": "^5.0.14",
     "@mui/material": "^5.16.7",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.26.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.9.0",
